requirements.txt
fastapi==0.115.3
uvicorn==0.30.1
watchfiles==0.21.0
httpx==0.28.1
python-dotenv==1.0.1
PyMuPDF==1.23.26
asyncpg==0.30.0
pgvector==0.4.1
tiktoken==0.5.2
